Add buildpaths class
The BuildPaths class builds up the command to run. It gets passed a hash
of options that is gathered by the Cli class, and then is used by the
Cli to execute terraform commands.

Need to add spec tests to ensure that the thing being reported back is
what we expect.