Redo BuildPaths class to enable testing
The way I previously wrote the BuildPaths class with lots of global
variables does not present itself well to testing because we want to
test against values that are returned by the methods themselves.

Instead of creating a bunch of global variables, return hashes, arrays
or values which can then test against when passing in dummy data.