Add "detailed-exitcode" as default to plan
This is a useful flag that should be included when we run a plan, so we
can use the exitcode to determine if an apply needs to happen upon a
project.

Update the error checking so we do not overwrite the exitcode, as the
abort writes the exitcode to 1 since it gathers that 2 is not a clean
exit.