Update error handling in config
The config loader is an unyeildy piece of code full of if statements,
and I struggle to make changes to tests without rspec aborting early,
but not raising any errors. I tried updating the exception, so the code
is a little nicer and will abort, and added specific errors into the
spec tests.

There is still a long way to go to make this more readable.