Allow app specific config to be loaded
This commit allows a user to specify app specific configuration that
takes precadence over "default" configuration in a config file, but can
still be overwritten by env vars or CLI flags.

The idea is to specify a default set of options:

```
default:
  environment: foo
  stack: bar
```

For a specific app, put the options under the name of the project you're
calling:

```
app-foo:
  stack: foo
```

This was slightly problematic to implement due to the cyclic nature of
ensuring the project name is already set when trying to find the
configuration under the project you're running against. I intend to
revisit how the configuration is loaded entirely as the number of
conditionals and ways of catching errors is complicated and extremely
difficult to understand.

There may also be some edge cases for when you specify a project under
the project specific configuration: in theory we should never get to
this point because the config would have loaded the config from the
default config file, env var or flag before we ever get there.

I also continue to have problems with my tests and exceptions, and in a
previous commit I moved to having an exception block, I backed this out
as I was finding it difficult to test properly. I need to also revisit
how to do exception handling properly (or find someone who can teach me
Ruby).