Add deployment method
Allows specifying a list of projects to deploy in turn using a
deployment file and specifying a command.

It was slightly more complicated than I intended due to the way that
configuration is loaded by configuration files, and so it meant that the
project name must always be loaded first.