Fix deployment bug
In the execution of Terraform code we change to the directory we're
going to run commands in. When running a single command this isn't
problematic, but when running through a loop the change of directory did
not reset itself which caused issues with paths.

This sets the current directory and then moves back to it after we're
done.