Refactor config into separate class
The CLI class is completely unyieldy and pretty much all the logic
exists in there, so this should be seperated out where possible and
other classes called.

Beginning with "config", where a lot of important logic takes place. I
have tidied up the code a little, and hopefully sanitised it so it is
easier to read.

Options take place as a hash so it's easier to pass stuff in, and easier
to test.

Cli class now has zero tests, so a mock should be created using the
"dryrun" option to see if we get the output we expect when passing in
different parameters.