Add end-to-end tests for CLI
This enables checking the output of what we expect to see from running
the command, so we can ensure that any changes don't have any impact on
the wrapper itself.

This is a first draft, and I hope a lot of complex and verbosity could
be wrapped up in a little logic.

It also changes some of the current tests due to the way we wish to test
the config file defaults.