Change dryrun to test
Update the flag "dryrun" to "test". This command is mostly used for when
running with tests, so make it more obvious that this is the case, so
that it is consistent when we use it across the codebase.

It also had a decrepency where the env var was "DRY_RUN" but the flag
was "dryrun", which I disliked.